Looked at tetris in a hex editor, and disassembler
Trying to follow Ferris, I looked at tetris in hecate
https://github.com/evanmiller/hecate

Each opcode I referenced in this awesome chart
http://www.pastraiser.com/cpu/gameboy/gameboy_opcodes.html

After googling for what supposed to happen in 0x100-0x103,
it supposed to be NOP, JMP 0x150. The hex editor showed
00 c3 50 01, which is NOP JP 01 50 - we have to switch bytes.
Not sure if the '8080/Z80' of the game boy is different
endianness than my laptop, but looks like it.

Searched for diasassemblers, and after finding 2 that don't
quite know about the game boy's weird variant, found
https://github.com/mmuszkow/gb-disasm which rules.