Implement beginning of memory mapping
This took like 4 hours, because I did not understand the example for the
Index trait. I thought the output of the Index trait required the same
signature as the `type Output`, so I was trying:

    type Output = u8;

    fn index<'a>(&'a self, index: u16) -> u8

I asked twice on #rust-beginners very specific questions about it, then
finally asked very generally "I don't understand the Index trait".
kimundi explained the signature should be more like

    fn index<'a>(&'a self, index: u16) -> &Self::Output

Which was much more helpful than the compiler errors that had me running
in circles before. I detailed my issues on github in hopes we can get
either better error messages, or I understand how to read the error
messages better

    https://github.com/rust-lang/rust/issues/33972

I also had trouble trying to make a hashmap from Range to references, so
for now I am just using a big if statement.

Now we are reading opcodes, and they are mapped to different references!

Onwards to implementing all the opcodes in the boot sequence, to draw
nintendo onto the screen.