Implement LDH (a8),A and POP HL
POP HL is the first command that gets run, I accidentally implementd
LDH (a8), A because I misread 0xE1 as 0xE0.

I am kind of confused about what happens for gameboy bootstrap.

Essentially, we are starting the program counter at 0x0100, which
gets mapped to byte 0 of the cart, which jumps us to 0x028B.

I think this means we are skipping the actual bootrom. Maybe I should
start the program counter at 0, and just implement the bootrom
instructions. It would feel real good to see the nintendo logo
rendered and scrolling down. Either way, now the stack looks
like it is working...