Refactor FLAG_, interconnect to memory, other stuff
Spent 3 days trying to make ldd_a_d8 generic. Fighting with
the borrow checker. Complete waste of time. Right now, just
going to have to copy-paste a bunch of very similar functions.

Tried speeding through ferris' tutorial to see how he does it
but it looks like most of the N64 commands are orthoganol to
each other. Or maybe I need to just switch to string mapping
to an array of registers.

I really don't have a good sense of how to make

    ldd_{a,f,b,c,d,e,h,l}_d8 just one function with 1 param