Fix CALL/RET, implement DEC_B, RLA, RL_C, INC_HL, LDI_HL_A, JR_Z
So there are still two decisions I haven't made yet, and its starting
to make it really hard to code. Do we save addresses in big-endian or
little-endian, and where should the stack pointer point to?

I think this is the simplest way:

                                          stack pointer
program counter  function       stack address    stack value
                                0xFFFF         > 0x00
                                0xFFFD           0x00
                                0xFFFE           0x00

         0x0028  CALL 0x0095

                                0xFFFF           0x28
                                0xFFFD           0x00
                                0xFFFE         > 0x00

         0x0095  RET

The reason I think it is simpler, is because we can implement CALL like:

      self.memory[self.sp] = self.pc & 0xFF;
      self.sp -= 1;
      self.memory[self.sp] = (self.pc >> 8) & 0xFF;
      self.sp -= 1;

If we try to keep the invariant that the stack pointer always points to
the next available memory cell,

  * update the stack pointer AFTER each write
  * update the stack pointer BEFORE each read

This means, whenever we read, we are less likely to make the assumption
of what the stack pointer is pointing to, and have to be more explicit.
I think. Maybe its a wash, but I'm glad that I am at least writing this
down to refer to later.

---

There are a lot of other small but important bugfixes I ran into, during
this, but now we have functions! Well, CALL and RET, which is awesome!

However, I am really starting to not like some things about the code:

 the inconsistency of variable and function names

   return_address_high vs addr_h
   self.hl() / self.set() / self.read_word()

   I am not sure why I keep switching between machine / datasheet language and human language

* copy-pasting functions with just one line changed

   See ldd_c_d8 ldd_a_d8, etc...

   This is because of how I setup my struct - it fights the borrow
   checker all the time. I've tried rewriting multiple times, creating
   Reg<T> wrapper types, but it just like, doesn't work at all. I am
   pretty close to just doing some ugly hack like
     ldd(reg_as_string: &'static str) {
       match reg_as_string { "reg_a" => reg_a = someval }
     }

* the operation pattern that repeats itself a bunch

    let size = 2; print_disassembly; CODE; size

This is a long commit message, maybe I should put these points as issues