Start implementing a GPU/LCD
Still a little unclear about the gpu in the gameboy.

Like, I am not sure if it is another chip, and if it runs
off of its own internal clock or a shared oscillator.

I need to read more. But I am pushing out this WIP
because I want to tweet something funny...

    // give me a bit of self.control ....
    if self.control & 0b01000000 {
      ...