Add intital LcdScreen
This took a while to make. At first, I tried just showing both the
DebugScreen and LcdScreen together, but got segfaults.

So I decided they would both implement the window::Drawable trait,
and I would just work with that. I had to Box the trait for it
to compile - I was trying to use generics, but I don't really
know how they work and how they are different.

Once again, #rust in freenode was helpful, especially
rkruppe, talchas, and c_wraith.

I also formatted with rustfmt, which is why this looks even bigger.
Now there is no segfault, but run --release is slow, and DEBUG=true
means the cpu does not actually run anymore, as DebugScreen::update
is the same as DebugScreen::debug before.

Todo

  * Thread DebugScreen::update instead of stopping the Cpu
  * Speedup LcdScreen::update for the cpu to work quicker