Share memory between window::Screen and the main thread
First, we remove ownership of the lcd by the cpu.

Then we use Arc<>, which is a pointer that knows how
many references there are to it.

Then, we clone the reference, and move ownership of the
cloned reference into the new window::Screen/lcd thread.

This does not work yet, because Arc::new() is an immutable
reference, or at least its being passed as immutable into
the new Cpu struct.

It might make more sense in the future, to revert this change,
since the the LCD actually does stall the cpu when drawing
on real game boy hardware, but for now, I want to see this async
especially for the debug screen.