Multithreading works!
The most important thing about this commit is that debugScreen and Cpu
have similar run() behaviour - loop forever, sleeping 1ms each loop,
and for the debugScreen, re-render every 16ms.

The next steps are to scope the read() and write() of shared memory
in their own blocks ({}) to reduce how long the locks are held.

I would also like to profile, but maybe I can switch to implementing the
lcd screen draw routines as well.