Stringly type ld_{afbcdehl}_{afbcdehl}
This does a few things, both good(+) and bad(-)

 + reduces the total line count
 + hides the fact that rust cannot self borrow like `0x78 => self.ld_a(self.b())`
 - makes use of strings instead of constants

In an ideal world, I would like to do:

  `0x78 => self.ld(&self.a, &self.b)`

But I am not sure how to do that. I could use macros to generate all the
variations of load functions, but part of my goal is that someone could
read this to learn game boy development, and macros may be distracting.